Rename `delete_place_by_guid` to `delete_all_visits_for_guid`.
This commit changes the behavior of `delete_place_by_guid` to only
delete the Place if it doesn't have any foreign key references, like
bookmarks, tags, or keywords. If it does, we can't delete the Place
outright, as that would cause a constraint violation at best, and
data loss at worst. Instead, we delete all visits for the Place,
but leave its row in `moz_places`, inserting tombstones if the Place
is synced.

This commit also renames `delete_place_by_guid` to clarify that it
might not actually delete the Place. However, the Kotlin method
retains its old name (`PlacesWriterConnection#deletePlace`) for now.

See #2520.